# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MoboClinic ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register User"
                ],
                "summary": "Register User",
                "description": "register user by first name, last name, phone number, email, password",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "phone",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "first_name",
                                        "example": "Segun"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "last_name",
                                        "example": "Aka"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "234894445"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record Created"
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/registerCompany": {
            "post": {
                "tags": [
                    "Register Company"
                ],
                "summary": "Register Company",
                "description": "register company by name, address, phone number, email, password",
                "operationId": "authRegisterCompany",
                "requestBody": {
                    "description": "Pass company credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "phone",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Excel Hospital"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "address",
                                        "example": "12 Abayo street"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "234894445"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record Created"
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/addTest": {
            "post": {
                "tags": [
                    "Add Test"
                ],
                "summary": "Add Test",
                "description": "Add Test by testName, price, description",
                "operationId": "addTest",
                "requestBody": {
                    "description": "Pass test credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "testName",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "testName": {
                                        "type": "string",
                                        "format": "testName",
                                        "example": "Malaria"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "price",
                                        "example": "4000"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description",
                                        "example": "A test for Malaria"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record Created"
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong testName, price, description provided, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tests": {
            "get": {
                "tags": [
                    "Get Tests"
                ],
                "summary": "Get Tests",
                "operationId": "tests",
                "responses": {
                    "200": {
                        "description": "Get all tests"
                    }
                }
            }
        },
        "/api/test/{id}": {
            "put": {
                "tags": [
                    "Update Test"
                ],
                "summary": "Update Test",
                "description": "Update the values of a test",
                "operationId": "updateTest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of test that is to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "testName",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "testName": {
                                        "type": "string",
                                        "format": "testName",
                                        "example": "Malaria"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "price",
                                        "example": "4000"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description",
                                        "example": "A test for Malaria"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid details supplied"
                    },
                    "404": {
                        "description": "Test not found"
                    },
                    "200": {
                        "description": "test updated"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete Test"
                ],
                "summary": "Delete Test",
                "description": "deletes a single test from the  tests",
                "operationId": "deleteTest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of test to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "test deleted"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key_security_example": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "api_key",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}